refactor to graph service
