+++ packages/pds/src/app-view/api/app/bsky/feed/getRepostedBy.ts
@@ -12,7 +12,7 @@ export default function (server: Server, ctx: AppContext) {
       const { services, db } = ctx
       const { ref } = db.db.dynamic
 
-      const actorService = ctx.services.appView.actor(ctx.db)
+      const graphService = ctx.services.appView.graph(ctx.db)
 
       let builder = db.db
         .selectFrom('repost')
@@ -25,7 +25,7 @@ export default function (server: Server, ctx: AppContext) {
         )
         .where(notSoftDeletedClause(ref('creator_repo')))
         .whereNotExists(
-          actorService.blockQb(requester, [ref('repost.creator')]),
+          graphService.blockQb(requester, [ref('repost.creator')]),
         )
         .selectAll('creator')
         .select(['repost.cid as cid', 'repost.createdAt as createdAt'])
