+++ packages/pds/src/app-view/api/app/bsky/graph/getFollowers.ts
@@ -14,6 +14,7 @@ export default function (server: Server, ctx: AppContext) {
       const { ref } = db.db.dynamic
 
       const actorService = services.appView.actor(db)
+      const graphService = services.appView.graph(db)
 
       const subjectRes = await actorService.getActor(actor)
       if (!subjectRes) {
@@ -31,7 +32,7 @@ export default function (server: Server, ctx: AppContext) {
         )
         .where(notSoftDeletedClause(ref('creator_repo')))
         .whereNotExists(
-          actorService.blockQb(requester, [ref('follow.subjectDid')]),
+          graphService.blockQb(requester, [ref('follow.subjectDid')]),
         )
         .selectAll('creator')
         .select(['follow.cid as cid', 'follow.createdAt as createdAt'])
