+++ packages/pds/src/app-view/api/app/bsky/graph/getLists.ts
@@ -13,24 +13,16 @@ export default function (server: Server, ctx: AppContext) {
       const { ref } = db.db.dynamic
 
       const actorService = services.appView.actor(db)
+      const graphService = services.appView.graph(db)
 
       const creatorRes = await actorService.getActor(actor)
       if (!creatorRes) {
         throw new InvalidRequestError(`Actor not found: ${actor}`)
       }
 
-      let listsReq = ctx.db.db
-        .selectFrom('list')
+      let listsReq = graphService
+        .getListsQb(requester)
         .where('list.creator', '=', creatorRes.did)
-        .selectAll('list')
-        .select(
-          ctx.db.db
-            .selectFrom('list_block')
-            .where('list_block.creator', '=', requester)
-            .whereRef('list_block.subjectUri', '=', ref('list.uri'))
-            .select('list_block.uri')
-            .as('viewerBlocked'),
-        )
 
       const keyset = new TimeCidKeyset(ref('list.indexedAt'), ref('list.cid'))
       listsReq = paginate(listsReq, {
@@ -43,24 +35,13 @@ export default function (server: Server, ctx: AppContext) {
         listsReq.execute(),
         actorService.views.profile(creatorRes, requester),
       ])
+      const profileMap = {
+        [creator.did]: creator,
+      }
 
-      const lists = listsRes.map((row) => ({
-        uri: row.uri,
-        creator: creator,
-        name: row.name,
-        purpose: row.purpose,
-        description: row.description ?? undefined,
-        descriptionFacets: row.descriptionFacets
-          ? JSON.parse(row.descriptionFacets)
-          : undefined,
-        avatar: row.avatarCid
-          ? ctx.imgUriBuilder.getCommonSignedUri('avatar', row.avatarCid)
-          : undefined,
-        indexedAt: row.indexedAt,
-        viewer: {
-          blocked: row.viewerBlocked ?? undefined,
-        },
-      }))
+      const lists = listsRes.map((row) =>
+        graphService.formatListView(row, profileMap),
+      )
 
       return {
         encoding: 'application/json',
