+++ packages/pds/src/app-view/services/actor/index.ts
@@ -1,10 +1,8 @@
 import Database from '../../../db'
 import { DidHandle } from '../../../db/tables/did-handle'
-import { DbRef, notSoftDeletedClause } from '../../../db/util'
+import { notSoftDeletedClause } from '../../../db/util'
 import { ActorViews } from './views'
 import { ImageUriBuilder } from '../../../image/uri'
-import { NotEmptyArray } from '@atproto/common'
-import { sql } from 'kysely'
 
 export class ActorService {
   constructor(public db: Database, public imgUriBuilder: ImageUriBuilder) {}
@@ -64,123 +62,6 @@ export class ActorService {
       const orderB = order[b.did] ?? order[b.handle.toLowerCase()]
       return orderA - orderB
     })
-  }
-
-  blockQb(requester: string, refs: NotEmptyArray<DbRef>) {
-    return this.actorBlockQb(requester, refs).union(
-      this.blockListQb(requester, refs),
-    )
-  }
-
-  actorBlockQb(requester: string, refs: NotEmptyArray<DbRef>) {
-    const subjectRefs = sql.join(refs)
-    return this.db.db
-      .selectFrom('actor_block')
-      .where((qb) =>
-        qb
-          .where('actor_block.creator', '=', requester)
-          .whereRef('actor_block.subjectDid', 'in', sql`(${subjectRefs})`),
-      )
-      .orWhere((qb) =>
-        qb
-          .where('actor_block.subjectDid', '=', requester)
-          .whereRef('actor_block.creator', 'in', sql`(${subjectRefs})`),
-      )
-      .select(['creator', 'subjectDid'])
-  }
-
-  blockListQb(requester: string, refs: NotEmptyArray<DbRef>) {
-    const subjectRefs = sql.join(refs)
-
-    return this.db.db
-      .selectFrom('list_block')
-      .innerJoin('list', 'list.uri', 'list_block.subjectUri')
-      .innerJoin('list_item', (join) =>
-        join
-          .onRef('list_item.creator', '=', 'list.creator')
-          .onRef('list_item.listUri', '=', 'list.uri'),
-      )
-      .where((qb) =>
-        qb
-          .where('list_block.creator', '=', requester)
-          .whereRef('list_item.subjectDid', 'in', sql`(${subjectRefs})`),
-      )
-      .orWhere((qb) =>
-        qb
-          .where('list_item.subjectDid', '=', requester)
-          .whereRef('list_block.creator', 'in', sql`(${subjectRefs})`),
-      )
-      .select([
-        'list_block.creator as creator',
-        'list_item.subjectDid as subjectDid',
-      ])
-  }
-
-  async getBlocks(
-    requester: string,
-    subjectHandleOrDid: string,
-  ): Promise<{ blocking: boolean; blockedBy: boolean }> {
-    let subjectDid
-    if (subjectHandleOrDid.startsWith('did:')) {
-      subjectDid = subjectHandleOrDid
-    } else {
-      const res = await this.db.db
-        .selectFrom('did_handle')
-        .where('handle', '=', subjectHandleOrDid)
-        .select('did')
-        .executeTakeFirst()
-      if (!res) {
-        return { blocking: false, blockedBy: false }
-      }
-      subjectDid = res.did
-    }
-
-    const accnts = [requester, subjectDid]
-    const actorBlockReq = this.db.db
-      .selectFrom('actor_block')
-      .where('creator', 'in', accnts)
-      .where('subjectDid', 'in', accnts)
-      .selectAll()
-
-    const listBlockReq = this.db.db
-      .selectFrom('list_block')
-      .innerJoin('list', 'list.uri', 'list_block.subjectUri')
-      .innerJoin('list_item', (join) =>
-        join
-          .onRef('list_item.creator', '=', 'list.creator')
-          .onRef('list_item.listUri', '=', 'list.uri'),
-      )
-      .where('list_block.creator', 'in', accnts)
-      .where('list_item.subjectDid', 'in', accnts)
-      .select([
-        'list_block.creator as creator',
-        'list_item.subjectDid as subjectDid',
-      ])
-
-    const [actorBlockRes, listBlockRes] = await Promise.all([
-      actorBlockReq.execute(),
-      listBlockReq.execute(),
-    ])
-
-    const blocking =
-      actorBlockRes.some(
-        (row) => row.creator === requester && row.subjectDid === subjectDid,
-      ) ||
-      listBlockRes.some(
-        (row) => row.creator === requester && row.subjectDid === subjectDid,
-      )
-    const blockedBy =
-      actorBlockRes.some(
-        (row) => row.creator === subjectDid && row.subjectDid === requester,
-      ) ||
-      listBlockRes.some(
-        (row) => row.creator === subjectDid && row.subjectDid === requester,
-      )
-
-    return {
-      blocking,
-      blockedBy,
-    }
   }
 }
